6.20 Move map state to HOC
